api: rename CPU pinning migration fields And add missing error checks And move all vcpu code to vcpu.go
