+++ pkg/virt-launcher/virtwrap/converter/vcpu/BUILD.bazel
@@ -2,11 +2,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "numa_placement.go",
-        "vcpu.go",
-        "vcpu_placement.go",
-    ],
+    srcs = ["vcpu.go"],
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter/vcpu",
     visibility = ["//visibility:public"],
     deps = [
