+++ staging/src/kubevirt.io/api/core/v1/deepcopy_generated.go
@@ -4101,8 +4101,8 @@ func (in *VirtualMachineInstanceMigrationState) DeepCopyInto(out *VirtualMachine
 		*out = new(MigrationConfiguration)
 		(*in).DeepCopyInto(*out)
 	}
-	if in.TargetNodeCPUSet != nil {
-		in, out := &in.TargetNodeCPUSet, &out.TargetNodeCPUSet
+	if in.TargetCPUSet != nil {
+		in, out := &in.TargetCPUSet, &out.TargetCPUSet
 		*out = make([]int, len(*in))
 		copy(*out, *in)
 	}
