+++ staging/src/kubevirt.io/api/core/v1/types.go
@@ -573,10 +573,10 @@ type VirtualMachineInstanceMigrationState struct {
 	// If the VMI requires dedicated CPUs, this field will
 	// hold the dedicated CPU set on the target node
 	// +listType=atomic
-	TargetNodeCPUSet []int `json:"targetNodeCPUSet,omitempty"`
+	TargetCPUSet []int `json:"targetCPUSet,omitempty"`
 	// If the VMI requires dedicated CPUs, this field will
 	// hold the numa topology on the target node
-	TargetTopology string `json:"targetTopology,omitempty"`
+	TargetNodeTopology string `json:"targetNodeTopology,omitempty"`
 }
 
 type MigrationAbortStatus string
